refactor(masterclass): the highlight vocabulary belongs to the exhibit
Era II needs the token kinds, the scope rule and the staleness hash, and none
of Era I's cyan classes, streaming slicer or dual-theme token shape. Sharing
through era1-playground would have inverted the dependency; sharing the whole
module would have forced one era to carry fields it never reads.

So only the vocabulary moves up. Token shapes stay era-local on purpose: Era I's
prefix carries a light/dark pair because its window follows the page theme,
Era II's editor carries one hex because it is a dark slab in both.

Era I re-exports what moved, so its own imports and tests are unchanged — and
its fingerprint still matches the committed tokens without regenerating, which
is what proves the flattening order survived the refactor.

Entire-Checkpoint: 5c7fef8c04ef
