+++ apps/web/app/[locale]/learn/masterclass-28-07-2026/demos/era1-playground/highlight/index.ts
@@ -8,9 +8,12 @@
  * Shiki must never be imported here: this module reaches the browser.
  */
 
+import { fingerprintText, type Kind } from "../../highlight";
 import type { PromptSeed } from "../selector";
 
-export type Kind = "comment" | "keyword" | "plain" | "punct" | "string";
+export type { Kind } from "../../highlight";
+// biome-ignore lint/performance/noBarrelFile: re-exports the shared vocabulary so this module's public surface stays unchanged
+export { kindFromScopes } from "../../highlight";
 
 export interface PrefixToken {
   /** Dark-theme hex. */
@@ -24,28 +27,6 @@ export interface PrefixToken {
 export interface CompletionToken {
   k: Kind;
   t: string;
-}
-
-/**
- * Priority order matters. Shiki returns `// a comment` as a single token whose
- * scope list ends in `punctuation.definition.comment.js`, so a last-scope-wins
- * rule would style every comment as punctuation.
- */
-export function kindFromScopes(scopes: string[]): Kind {
-  const joined = scopes.join(" ");
-  if (joined.includes("comment")) {
-    return "comment";
-  }
-  if (joined.includes("string")) {
-    return "string";
-  }
-  if (joined.includes("storage.type") || joined.includes("keyword")) {
-    return "keyword";
-  }
-  if (joined.includes("punctuation")) {
-    return "punct";
-  }
-  return "plain";
 }
 
 const COMPLETION_CLASS: Record<Kind, string> = {
@@ -112,13 +93,5 @@ export function fingerprintSources(prompts: readonly PromptSeed[]): string {
       parts.push(p.continuations[band], p.instructAnswers[band]);
     }
   }
-  const joined = parts.join("");
-  let h1 = 0x81_1c_9d_c5;
-  for (let i = 0; i < joined.length; i += 1) {
-    // biome-ignore lint/suspicious/noBitwiseOperators: FNV-1a hash requires XOR and unsigned-shift
-    h1 ^= joined.charCodeAt(i);
-    // biome-ignore lint/suspicious/noBitwiseOperators: FNV-1a hash requires XOR and unsigned-shift
-    h1 = Math.imul(h1, 0x01_00_01_93) >>> 0;
-  }
-  return `${h1.toString(16).padStart(8, "0")}-${joined.length.toString(16)}`;
+  return fingerprintText(parts);
 }
